Try moving the test client
