+++ server/blobs_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/robinkb/cascade-registry/server"
 	"github.com/robinkb/cascade-registry/store"
 	. "github.com/robinkb/cascade-registry/testing"
+	testclient "github.com/robinkb/cascade-registry/testing/client"
 	"github.com/robinkb/cascade-registry/testing/mock"
 )
 
@@ -24,7 +25,7 @@ func TestStatBlob(t *testing.T) {
 			StatBlob(name, digest.String()).
 			Return(&store.BlobInfo{Size: size}, nil)
 
-		client := NewTestClientForRepository(t, name, repo)
+		client := testclient.NewTestClientForRepository(t, name, repo)
 
 		resp := client.CheckBlob(name, digest)
 
@@ -39,7 +40,7 @@ func TestStatBlob(t *testing.T) {
 			StatBlob(name, digest.String()).
 			Return(nil, repository.ErrBlobUnknown)
 
-		client := NewTestClientForRepository(t, name, repo)
+		client := testclient.NewTestClientForRepository(t, name, repo)
 
 		resp := client.CheckBlob(name, digest)
 
@@ -52,7 +53,7 @@ func TestStatBlob(t *testing.T) {
 			GetRepository(name).
 			Return(nil, repository.ErrNameUnknown)
 
-		client := NewTestClientForHandler(t, server.New(registry))
+		client := testclient.NewTestClientForHandler(t, server.New(registry))
 		resp := client.CheckBlob(name, digest)
 		AssertResponseCode(t, resp, http.StatusNotFound)
 	})
@@ -68,7 +69,7 @@ func TestGetBlob(t *testing.T) {
 			GetBlob(name, digest.String()).
 			Return(bytes.NewBuffer(content), nil)
 
-		client := NewTestClientForRepository(t, name, repo)
+		client := testclient.NewTestClientForRepository(t, name, repo)
 
 		resp := client.GetBlob(name, digest)
 
@@ -82,7 +83,7 @@ func TestGetBlob(t *testing.T) {
 			GetBlob(name, digest.String()).
 			Return(nil, repository.ErrBlobUnknown)
 
-		client := NewTestClientForRepository(t, name, repo)
+		client := testclient.NewTestClientForRepository(t, name, repo)
 
 		resp := client.GetBlob(name, digest)
 
@@ -96,7 +97,7 @@ func TestGetBlob(t *testing.T) {
 			GetRepository(name).
 			Return(nil, repository.ErrNameUnknown)
 
-		client := NewTestClientForHandler(t, server.New(registry))
+		client := testclient.NewTestClientForHandler(t, server.New(registry))
 		resp := client.GetBlob(name, digest)
 		AssertResponseCode(t, resp, http.StatusNotFound)
 	})
@@ -111,7 +112,7 @@ func TestDeleteBlob(t *testing.T) {
 			DeleteBlob(name, digest.String()).
 			Return(nil)
 
-		client := NewTestClientForRepository(t, name, repo)
+		client := testclient.NewTestClientForRepository(t, name, repo)
 
 		resp := client.DeleteBlob(name, digest)
 
@@ -124,7 +125,7 @@ func TestDeleteBlob(t *testing.T) {
 			GetRepository(name).
 			Return(nil, repository.ErrNameUnknown)
 
-		client := NewTestClientForHandler(t, server.New(registry))
+		client := testclient.NewTestClientForHandler(t, server.New(registry))
 		resp := client.DeleteBlob(name, digest)
 		AssertResponseCode(t, resp, http.StatusNotFound)
 	})
@@ -132,7 +133,7 @@ func TestDeleteBlob(t *testing.T) {
 
 func TestBlobsOthers(t *testing.T) {
 	t.Run("other methods return 405", func(t *testing.T) {
-		client := NewTestClientForHandler(t, server.New(nil))
+		client := testclient.NewTestClientForHandler(t, server.New(nil))
 
 		resp := client.Do(http.MethodConnect, "/v2/library/fedora/blobs/123", nil, nil)
