+++ testing/conformance/push_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/robinkb/cascade-registry/server"
 	"github.com/robinkb/cascade-registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
+	testclient "github.com/robinkb/cascade-registry/testing/client"
 )
 
 func TestPush(t *testing.T) {
@@ -22,7 +23,7 @@ func TestPush(t *testing.T) {
 	t.Run("Pushing blobs", func(t *testing.T) {
 		t.Run("Pushing a blob monolithically", func(t *testing.T) {
 			t.Run("POST then PUT", func(t *testing.T) {
-				client := NewTestClientForHandler(t, srv)
+				client := testclient.NewTestClientForHandler(t, srv)
 
 				name := RandomName()
 				digest, blob := RandomBlob(32)
@@ -49,7 +50,7 @@ func TestPush(t *testing.T) {
 			})
 
 			t.Run("Single POST", func(t *testing.T) {
-				client := NewTestClientForHandler(t, srv)
+				client := testclient.NewTestClientForHandler(t, srv)
 
 				// Registries MAY support pushing blobs using a single POST request.
 				// Cascade does not.
@@ -66,7 +67,7 @@ func TestPush(t *testing.T) {
 		})
 
 		t.Run("Pushing a blob in chunks", func(t *testing.T) {
-			client := NewTestClientForHandler(t, srv)
+			client := testclient.NewTestClientForHandler(t, srv)
 
 			name := RandomName()
 			digest, blob := RandomBlob(64 * 1024)
@@ -141,7 +142,7 @@ func TestPush(t *testing.T) {
 
 		// This is not part of the spec (yet).
 		t.Run("Pushing a blob as a stream", func(t *testing.T) {
-			client := NewTestClientForHandler(t, srv)
+			client := testclient.NewTestClientForHandler(t, srv)
 
 			name := RandomName()
 			digest, blob := RandomBlob(64 * 1024)
@@ -167,7 +168,7 @@ func TestPush(t *testing.T) {
 
 	t.Run("Pushing manifests", func(t *testing.T) {
 		t.Run("Pushing manifest by digest", func(t *testing.T) {
-			client := NewTestClientForHandler(t, srv)
+			client := testclient.NewTestClientForHandler(t, srv)
 
 			name := RandomName()
 			digest, _, content := RandomManifest()
@@ -186,7 +187,7 @@ func TestPush(t *testing.T) {
 		})
 
 		t.Run("Pushing manifest by tag", func(t *testing.T) {
-			client := NewTestClientForHandler(t, srv)
+			client := testclient.NewTestClientForHandler(t, srv)
 
 			name := RandomName()
 			digest, _, content := RandomManifest()
@@ -212,7 +213,7 @@ func TestPush(t *testing.T) {
 		t.Run("Pushing manifest with layers", func(t *testing.T) {})
 
 		t.Run("Pushing manifests with Subject", func(t *testing.T) {
-			client := NewTestClientForHandler(t, srv)
+			client := testclient.NewTestClientForHandler(t, srv)
 
 			name := RandomName()
 			subjectDigest, subjectManifest, _ := RandomManifest()
